fix(router/jlu): add parameters and features
